refactor(master): Use std::vector for chunk buckets
This commit replaces the C-linked list of chunk_bucket structs with
a std::vector of smart pointers to ChunkBucket. This simplifies memory
management and improves code clarity.

Additionally, it was applied a small refactoring to rename some
variables with more meaningful names and remove redundant void and
return statements from void functions.

Signed-off-by: Crash <crash@leil.io>